Split transport job tests below module gate
